update arg parse

+++ models/swin_transformer.py
@@ -49,11 +49,14 @@ def create_mask(window_size, displacement, upper_lower, left_right):
     mask = torch.zeros(window_size ** 2, window_size ** 2)
 
     if upper_lower:
-        mask[-displacement * window_size:, :-displacement * window_size] = float('-inf')
-        mask[:-displacement * window_size, -displacement * window_size:] = float('-inf')
+        mask[-displacement * window_size:, :-
+             displacement * window_size] = float('-inf')
+        mask[:-displacement * window_size, -
+             displacement * window_size:] = float('-inf')
 
     if left_right:
-        mask = rearrange(mask, '(h1 w1) (h2 w2) -> h1 w1 h2 w2', h1=window_size, h2=window_size)
+        mask = rearrange(mask, '(h1 w1) (h2 w2) -> h1 w1 h2 w2',
+                         h1=window_size, h2=window_size)
         mask[:, -displacement:, :, :-displacement] = float('-inf')
         mask[:, :-displacement, :, -displacement:] = float('-inf')
         mask = rearrange(mask, 'h1 w1 h2 w2 -> (h1 w1) (h2 w2)')
@@ -62,7 +65,8 @@ def create_mask(window_size, displacement, upper_lower, left_right):
 
 
 def get_relative_distances(window_size):
-    indices = torch.tensor(np.array([[x, y] for x in range(window_size) for y in range(window_size)]))
+    indices = torch.tensor(
+        np.array([[x, y] for x in range(window_size) for y in range(window_size)]))
     distances = indices[None, :, :] - indices[:, None, :]
     return distances
 
@@ -90,10 +94,13 @@ class WindowAttention(nn.Module):
         self.to_qkv = nn.Linear(dim, inner_dim * 3, bias=False)
 
         if self.relative_pos_embedding:
-            self.relative_indices = get_relative_distances(window_size) + window_size - 1
-            self.pos_embedding = nn.Parameter(torch.randn(2 * window_size - 1, 2 * window_size - 1))
+            self.relative_indices = get_relative_distances(
+                window_size) + window_size - 1
+            self.pos_embedding = nn.Parameter(torch.randn(
+                2 * window_size - 1, 2 * window_size - 1))
         else:
-            self.pos_embedding = nn.Parameter(torch.randn(window_size ** 2, window_size ** 2))
+            self.pos_embedding = nn.Parameter(
+                torch.randn(window_size ** 2, window_size ** 2))
 
         self.to_out = nn.Linear(inner_dim, dim)
 
@@ -114,7 +121,8 @@ class WindowAttention(nn.Module):
         dots = einsum('b h w i d, b h w j d -> b h w i j', q, k) * self.scale
 
         if self.relative_pos_embedding:
-            dots += self.pos_embedding[self.relative_indices[:, :, 0], self.relative_indices[:, :, 1]]
+            dots += self.pos_embedding[self.relative_indices[:, :, 0].type(
+                torch.long), self.relative_indices[:, :, 1].type(torch.long)]
         else:
             dots += self.pos_embedding
 
@@ -143,7 +151,8 @@ class SwinBlock(nn.Module):
                                                                      shifted=shifted,
                                                                      window_size=window_size,
                                                                      relative_pos_embedding=relative_pos_embedding)))
-        self.mlp_block = Residual(PreNorm(dim, FeedForward(dim=dim, hidden_dim=mlp_dim)))
+        self.mlp_block = Residual(
+            PreNorm(dim, FeedForward(dim=dim, hidden_dim=mlp_dim)))
 
     def forward(self, x):
         x = self.attention_block(x)
@@ -155,8 +164,10 @@ class PatchMerging(nn.Module):
     def __init__(self, in_channels, out_channels, downscaling_factor):
         super().__init__()
         self.downscaling_factor = downscaling_factor
-        self.patch_merge = nn.Unfold(kernel_size=downscaling_factor, stride=downscaling_factor, padding=0)
-        self.linear = nn.Linear(in_channels * downscaling_factor ** 2, out_channels)
+        self.patch_merge = nn.Unfold(
+            kernel_size=downscaling_factor, stride=downscaling_factor, padding=0)
+        self.linear = nn.Linear(
+            in_channels * downscaling_factor ** 2, out_channels)
 
     def forward(self, x):
         b, c, h, w = x.shape
